backend-api/requirements.txt
fastapi>=0.115.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.9
llama-index>=0.12.0
llama-index-llms-ollama>=0.5.0
llama-index-embeddings-huggingface>=0.5.0
llama-index-postprocessor-sbert-rerank>=0.3.0
llama-index-readers-file>=0.4.0
llama-index-vector-stores-qdrant>=0.4.0
qdrant-client>=1.9.0
llama-index-graph-stores-neo4j>=0.3.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
pymupdf>=1.24.0
pandas>=2.0.0
pytesseract>=0.3.10
pillow>=10.0.0
